+++ tests/drum/test_fit.py
@@ -58,6 +58,8 @@ from .constants import (
     SKLEARN_BINARY_HYPERPARAMETERS,
     SKLEARN_TRANSFORM_HYPERPARAMETERS,
     SKLEARN_TRANSFORM_PARAMETERS,
+    RDS_HYPERPARAMETERS,
+    RDS_PARAMETERS,
 )
 
 
@@ -209,13 +211,14 @@ class TestFit:
             (SKLEARN_BINARY_HYPERPARAMETERS, BINARY_TEXT, None, SKLEARN_BINARY_PARAMETERS),
             (SKLEARN_BINARY_HYPERPARAMETERS, BINARY_SPACES, None, SKLEARN_BINARY_PARAMETERS),
             (SKLEARN_TRANSFORM_HYPERPARAMETERS, REGRESSION, None, SKLEARN_TRANSFORM_PARAMETERS),
+            (RDS_HYPERPARAMETERS, BINARY_TEXT, None, RDS_PARAMETERS),
         ],
     )
     @pytest.mark.parametrize("weights", [WEIGHTS_CSV, WEIGHTS_ARGS, None])
     def test_fit_hyperparameters(
         self, resources, framework, problem, docker, parameters, weights, tmp_path,
     ):
-        if framework == RDS:
+        if framework == RDS_HYPERPARAMETERS:
             language = R_FIT
         else:
             language = PYTHON